+++ src/main/java/cmu/pasta/mu2/fuzz/DeadMutantsFilter.java
@@ -21,8 +21,4 @@ public class DeadMutantsFilter implements MutantFilter {
         return aliveMuts;
     }
 
-    @Override
-    public void prepTrial() {
-        return;
-    }
 }
